+++ database/migrations/2024_02_02_053055_drop_price_column_in_plans_table.php
@@ -13,7 +13,9 @@ return new class extends Migration
     {
         Schema::table('plans', function (Blueprint $table) {
             //
-            $table->dropColumn('price');
+            if (Schema::hasColumn('plans', 'price')) {
+                $table->dropColumn('price');
+            }
         });
     }
 
@@ -23,7 +25,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('plans', function (Blueprint $table) {
-            $table->decimal('price', 10, 2);
+            if (!Schema::hasColumn('plans', 'price')) {
+                $table->decimal('price', 10, 2);
+            }
         });
     }
 };
