+++ database/migrations/2024_02_21_111850_add_status_column_in_users_table.php
@@ -13,7 +13,9 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             //
-            $table->tinyInteger('status')->default(0)->after('id');
+            if (!Schema::hasColumn('users', 'status')) {
+                $table->tinyInteger('status')->default(0)->after('id');
+            }
         });
     }
 
@@ -24,7 +26,9 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             //
-            $table->dropColumn('status');
+            if (Schema::hasColumn('users', 'status')) {
+                $table->dropColumn('status');
+            }
         });
     }
 };
