+++ database/migrations/2024_02_27_061558_add_column_reset_token_in_users_table.php
@@ -13,7 +13,9 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             //
-            $table->string('reset_token')->nullable();
+            if (!Schema::hasColumn('users', 'reset_token')) {
+                $table->string('reset_token')->nullable();
+            }
         });
     }
 
@@ -24,7 +26,9 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             //
-            $table->dropColumn('reset_token');
+            if (Schema::hasColumn('users', 'reset_token')) {
+                $table->dropColumn('reset_token');
+            }
         });
     }
 };
